requirements.txt
flask
requests
upstash_redis
python-dotenv